Allow the calls.pkl path to be explicitly specified.
In order to allow for hermetic builds that don't touch files outside of those explicitly provided, this allows a `System` to have its models files specified explicitly. This allows portability between machines, where models can be pre-compiled into images for use in server environments, and unit tests can be made hermetic. Additionally, this avoids race conditions to modify the same calls.pkl file with multiple threads.

Ideally, a refactor of the paths module into an injected dependency would make this more clean, but that will take a bit more time. Perhaps a models-store object that is injected would be a better option for a future refactor.